refactor(utils): simplify exporting utils public APIs
