+++ src/cli/cli.spec.ts
@@ -1,7 +1,7 @@
 import * as _fs from 'fs'
 import { normalize, resolve } from 'path'
 
-import { mocked } from '../utils'
+import { mocked } from '../utils/testing'
 import { logTargetMock, mockObject, mockWriteStream } from '../__helpers__/mocks'
 
 import { processArgv } from '.'
