--- src/utils/index.ts
@@ -1,3 +1,0 @@
-export { mocked } from './testing'
-export { createJestPreset } from '../presets/create-jest-preset'
-export { pathsToModuleNameMapper } from '../config/paths-to-module-name-mapper'
