+++ utils/index.d.ts
@@ -1,1 +1,3 @@
-export * from '../dist/utils'
+export { pathsToModuleNameMapper } from '../dist/config/paths-to-module-name-mapper'
+export { createJestPreset } from '../dist/presets/create-jest-preset'
+export { mocked } from '../dist/utils/testing'
